+++ src/components/VCombobox/index.js
@@ -1,0 +1,9 @@
+import VCombobox from './VCombobox'
+
+/* istanbul ignore next */
+VCombobox.install = function install (Vue) {
+  Vue.component(VCombobox.name, VCombobox)
+}
+
+export { VCombobox }
+export default VCombobox
