+++ src/components/VSelect/index.js
@@ -1,6 +1,7 @@
 import VSelect from './VSelect'
 import VOverflowBtn from '../VOverflowBtn'
 import VAutocomplete from '../VAutocomplete'
+import VCombobox from '../VCombobox'
 import rebuildSlots from '../../util/rebuildFunctionalSlots'
 import { deprecate } from '../../util/console'
 
@@ -30,10 +31,10 @@ const wrapper = {
       deprecate('<v-select autocomplete>', '<v-autocomplete>', wrapper, parent)
     }
     if (props.combobox) {
-      deprecate('<v-select combobox>', '<v-autocomplete combobox>', wrapper, parent)
+      deprecate('<v-select combobox>', '<v-combobox>', wrapper, parent)
     }
     if (props.tags) {
-      deprecate('<v-select tags>', '<v-autocomplete combobox multiple>', wrapper, parent)
+      deprecate('<v-select tags>', '<v-combobox multiple>', wrapper, parent)
     }
 
     if (props.overflow) {
@@ -46,9 +47,11 @@ const wrapper = {
       deprecate('<v-select editable>', '<v-overflow-btn editable>', wrapper, parent)
     }
 
-    if (props.autocomplete || props.combobox || props.tags) {
-      data.attrs.combobox = props.combobox || props.tags
-      data.attrs.multiple = props.multiple || props.tags
+    if (props.combobox || props.tags) {
+      data.attrs.multiple = props.tags
+      return h(VCombobox, data, children)
+    } else if (props.autocomplete) {
+      data.attrs.multiple = props.multiple
       return h(VAutocomplete, data, children)
     } else if (props.overflow || props.segmented || props.editable) {
       data.attrs.segmented = props.segmented
