+++ src/components/index.js
@@ -13,6 +13,7 @@ export { default as VCard } from './VCard'
 export { default as VCarousel } from './VCarousel'
 export { default as VCheckbox } from './VCheckbox'
 export { default as VChip } from './VChip'
+export { default as VCombobox } from './VCombobox'
 export { default as VCounter } from './VCounter'
 export { default as VDataIterator } from './VDataIterator'
 export { default as VDataTable } from './VDataTable'
